+++ packages/analyzer_kit_annotation/example/analyzer_kit_annotation_example.dart
@@ -1,0 +1,18 @@
+import 'package:analyzer_kit_annotation/analyzer_kit_annotation.dart';
+
+@CopyWith()
+final class User {
+  const User({required this.name, required this.age});
+
+  final String name;
+  final int age;
+}
+
+@copyWith
+final class Todo {
+  const Todo({required this.id, required this.title, required this.completed});
+
+  final int id;
+  final String title;
+  final bool completed;
+}