+++ packages/analyzer_kit_annotation/lib/analyzer_kit_annotation.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'src/copy_with.dart';