+++ packages/analyzer_kit_annotation/test/analyzer_kit_annotation_test.dart
@@ -1,0 +1,13 @@
+import 'package:analyzer_kit_annotation/analyzer_kit_annotation.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('copyWith', () {
+    final instance = CopyWith();
+    final constant = copyWith;
+
+    expect(constant, isNotNull);
+    expect(instance, isNotNull);
+    expect(constant, equals(instance));
+  });
+}