+++ packages/dart_analyzer_kit/CHANGELOG.md
@@ -3,3 +3,6 @@
 - Initial version.
 - Added `unused_copy_with_annotation` lint rule.
 - Added `add_copy_with_method` quick fix.
+- Removed `CopyWith` annotation from the library.
+    - Now `CopyWith` is a part of `analyzer_kit_annotation` package.
+    - Rest of the annotations will be kept separately in `analyzer_kit_annotations` package.